Consolidate REE and RPMB filesystems
The code is re-organized so that we have fewer C files, all in the
core/tee folder:
 - tee_svc_storage.c contains the syscalls that implement the GP API.
   It depends on tee_ree_fs.c or tee_rpmb_fs.c to export a
   tee_file_operations structure. FS-specific code is entirely in
   static functions in these files.
 - tee_ree_fs.c is the REE filesystem code
 - tee_rpmb_fs.c is the RPMB filesystem code
 - tee_fs_key_manager.c provides block encryption and key management
