+++ core/arch/arm/tee/sub.mk
@@ -1,6 +1,4 @@
 ifeq ($(CFG_WITH_USER_TA),y)
-srcs-y += arch_tee_fs.c
-srcs-y += tee_rpmb.c
 srcs-$(CFG_ARM32_core) += arch_svc_a32.S
 srcs-$(CFG_ARM64_core) += arch_svc_a64.S
 srcs-$(CFG_CACHE_API) += svc_cache.c
