+++ core/tee/tee_rpmb_fs.c
@@ -25,24 +25,25 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <assert.h>
 #include <kernel/tee_common.h>
 #include <kernel/handle.h>
-#include <tee/tee_rpmb_fs.h>
-#include <tee/tee_rpmb.h>
+#include <kernel/mutex.h>
+#include <kernel/tee_common_otp.h>
+#include <kernel/thread.h>
+#include <optee_msg.h>
+#include <tee/tee_cryp_provider.h>
 #include <tee/tee_fs_defs.h>
 #include <tee/tee_fs.h>
 #include <tee/tee_fs_key_manager.h>
+#include <mm/core_memprot.h>
 #include <mm/tee_mm.h>
-#include <trace.h>
 #include <stdlib.h>
 #include <string.h>
 #include <string_ext.h>
-#include <util.h>
 #include <sys/queue.h>
-
-#ifdef CFG_ENC_FS
-#include <tee/tee_cryp_provider.h>
-#endif
+#include <trace.h>
+#include <util.h>
 
 #define RPMB_STORAGE_START_ADDRESS      0
 #define RPMB_FS_FAT_START_ADDRESS       512
@@ -54,6 +55,13 @@
 
 #define FILE_IS_ACTIVE                  (1u << 0)
 #define FILE_IS_LAST_ENTRY              (1u << 1)
+
+#define TEE_RPMB_FS_FILENAME_LENGTH 224
+
+struct tee_rpmb_fs_stat {
+	size_t size;
+	uint32_t reserved;
+};
 
 /**
  * FS parameters: Information often used by internal functions.
@@ -82,14 +90,13 @@ struct rpmb_fat_entry {
  * location in RPMB.
  */
 struct rpmb_file_handle {
-	/* Pointer to a fat_entry */
 	struct rpmb_fat_entry fat_entry;
-	/* Pointer to a filename */
 	char filename[TEE_RPMB_FS_FILENAME_LENGTH];
-	/* Adress for current entry in RPMB */
+	/* Address for current entry in RPMB */
 	uint32_t rpmb_fat_address;
 	/* Current position */
 	uint32_t pos;
+	uint32_t flags;
 };
 
 /**
@@ -111,6 +118,7 @@ struct rpmb_fs_partition {
 struct tee_rpmb_fs_dirent {
 	struct tee_fs_dirent entry;
 	char name[TEE_RPMB_FS_FILENAME_LENGTH];
+	/* */
 	SIMPLEQ_ENTRY(tee_rpmb_fs_dirent) link;
 };
 
@@ -122,14 +130,1430 @@ struct tee_rpmb_fs_dirent {
  */
 struct tee_fs_dir {
 	struct tee_rpmb_fs_dirent *current;
+	/* */
 	SIMPLEQ_HEAD(next_head, tee_rpmb_fs_dirent) next;
 };
-
-static TEE_Result get_fat_start_address(uint32_t *addr);
 
 static struct rpmb_fs_parameters *fs_par;
 
 static struct handle_db fs_handle_db = HANDLE_DB_INITIALIZER;
+
+/*
+ * Lower interface to RPMB device
+ */
+
+#define RPMB_DATA_OFFSET            (RPMB_STUFF_DATA_SIZE + RPMB_KEY_MAC_SIZE)
+#define RPMB_MAC_PROTECT_DATA_SIZE  (RPMB_DATA_FRAME_SIZE - RPMB_DATA_OFFSET)
+
+#define RPMB_MSG_TYPE_REQ_AUTH_KEY_PROGRAM          0x0001
+#define RPMB_MSG_TYPE_REQ_WRITE_COUNTER_VAL_READ    0x0002
+#define RPMB_MSG_TYPE_REQ_AUTH_DATA_WRITE           0x0003
+#define RPMB_MSG_TYPE_REQ_AUTH_DATA_READ            0x0004
+#define RPMB_MSG_TYPE_REQ_RESULT_READ               0x0005
+#define RPMB_MSG_TYPE_RESP_AUTH_KEY_PROGRAM         0x0100
+#define RPMB_MSG_TYPE_RESP_WRITE_COUNTER_VAL_READ   0x0200
+#define RPMB_MSG_TYPE_RESP_AUTH_DATA_WRITE          0x0300
+#define RPMB_MSG_TYPE_RESP_AUTH_DATA_READ           0x0400
+
+#define RPMB_STUFF_DATA_SIZE                        196
+#define RPMB_KEY_MAC_SIZE                           32
+#define RPMB_DATA_SIZE                              256
+#define RPMB_NONCE_SIZE                             16
+#define RPMB_DATA_FRAME_SIZE                        512
+
+#define RPMB_RESULT_OK                              0x00
+#define RPMB_RESULT_GENERAL_FAILURE                 0x01
+#define RPMB_RESULT_AUTH_FAILURE                    0x02
+#define RPMB_RESULT_COUNTER_FAILURE                 0x03
+#define RPMB_RESULT_ADDRESS_FAILURE                 0x04
+#define RPMB_RESULT_WRITE_FAILURE                   0x05
+#define RPMB_RESULT_READ_FAILURE                    0x06
+#define RPMB_RESULT_AUTH_KEY_NOT_PROGRAMMED         0x07
+#define RPMB_RESULT_MASK                            0x3F
+#define RPMB_RESULT_WR_CNT_EXPIRED                  0x80
+
+/* RPMB internal commands */
+#define RPMB_CMD_DATA_REQ      0x00
+#define RPMB_CMD_GET_DEV_INFO  0x01
+
+#define RPMB_SIZE_SINGLE (128 * 1024)
+
+/* Error codes for get_dev_info request/response. */
+#define RPMB_CMD_GET_DEV_INFO_RET_OK     0x00
+#define RPMB_CMD_GET_DEV_INFO_RET_ERROR  0x01
+
+struct rpmb_data_frame {
+	uint8_t stuff_bytes[RPMB_STUFF_DATA_SIZE];
+	uint8_t key_mac[RPMB_KEY_MAC_SIZE];
+	uint8_t data[RPMB_DATA_SIZE];
+	uint8_t nonce[RPMB_NONCE_SIZE];
+	uint8_t write_counter[4];
+	uint8_t address[2];
+	uint8_t block_count[2];
+	uint8_t op_result[2];
+	uint8_t msg_type[2];
+};
+
+struct rpmb_req {
+	uint16_t cmd;
+	uint16_t dev_id;
+	uint16_t block_count;
+	/* variable length of data */
+	/* uint8_t data[]; REMOVED! */
+};
+
+#define TEE_RPMB_REQ_DATA(req) \
+		((void *)((struct rpmb_req *)(req) + 1))
+
+struct rpmb_raw_data {
+	uint16_t msg_type;
+	uint16_t *op_result;
+	uint16_t *block_count;
+	uint16_t *blk_idx;
+	uint32_t *write_counter;
+	uint8_t *nonce;
+	uint8_t *key_mac;
+	uint8_t *data;
+	/* data length to read or write */
+	uint32_t len;
+	/* Byte address offset in the first block involved */
+	uint8_t byte_offset;
+};
+
+#define RPMB_EMMC_CID_SIZE 16
+struct rpmb_dev_info {
+	uint8_t cid[RPMB_EMMC_CID_SIZE];
+	/* EXT CSD-slice 168 "RPMB Size" */
+	uint8_t rpmb_size_mult;
+	/* EXT CSD-slice 222 "Reliable Write Sector Count" */
+	uint8_t rel_wr_sec_c;
+	/* Check the ret code and accept the data only if it is OK. */
+	uint8_t ret_code;
+};
+
+/*
+ * Struct for rpmb context data.
+ *
+ * @key              RPMB key.
+ * @cid              eMMC card ID.
+ * @hash_ctx_size    Hash context size
+ * @wr_cnt           Current write counter.
+ * @max_blk_idx      The highest block index supported by current device.
+ * @rel_wr_blkcnt    Max number of data blocks for each reliable write.
+ * @dev_id           Device ID of the eMMC device.
+ * @wr_cnt_synced    Flag indicating if write counter is synced to RPMB.
+ * @key_derived      Flag indicating if key has been generated.
+ * @key_verified     Flag indicating the key generated is verified ok.
+ * @dev_info_synced  Flag indicating if dev info has been retrieved from RPMB.
+ */
+struct tee_rpmb_ctx {
+	uint8_t key[RPMB_KEY_MAC_SIZE];
+	uint8_t cid[RPMB_EMMC_CID_SIZE];
+	size_t hash_ctx_size;
+	uint32_t wr_cnt;
+	uint16_t max_blk_idx;
+	uint16_t rel_wr_blkcnt;
+	uint16_t dev_id;
+	bool wr_cnt_synced;
+	bool key_derived;
+	bool key_verified;
+	bool dev_info_synced;
+};
+
+static struct tee_rpmb_ctx *rpmb_ctx;
+
+/*
+ * Mutex to serialize the operations exported by this file.
+ * It protects rpmb_ctx and prevents overlapping operations on eMMC devices with
+ * different IDs.
+ */
+static struct mutex rpmb_mutex = MUTEX_INITIALIZER;
+
+#ifdef CFG_RPMB_TESTKEY
+
+static const uint8_t rpmb_test_key[RPMB_KEY_MAC_SIZE] = {
+	0xD3, 0xEB, 0x3E, 0xC3, 0x6E, 0x33, 0x4C, 0x9F,
+	0x98, 0x8C, 0xE2, 0xC0, 0xB8, 0x59, 0x54, 0x61,
+	0x0D, 0x2B, 0xCF, 0x86, 0x64, 0x84, 0x4D, 0xF2,
+	0xAB, 0x56, 0xE6, 0xC6, 0x1B, 0xB7, 0x01, 0xE4
+};
+
+static TEE_Result tee_rpmb_key_gen(uint16_t dev_id __unused,
+				   uint8_t *key, uint32_t len)
+{
+	TEE_Result res = TEE_SUCCESS;
+
+	if (!key || RPMB_KEY_MAC_SIZE != len) {
+		res = TEE_ERROR_BAD_PARAMETERS;
+		goto out;
+	}
+
+	DMSG("RPMB: Using test key");
+	memcpy(key, rpmb_test_key, RPMB_KEY_MAC_SIZE);
+
+out:
+	return res;
+}
+
+#else /* !CFG_RPMB_TESTKEY */
+
+/*
+ * NOTE: We need a common API to get hw unique key and it
+ * should return error when the hw unique is not a valid
+ * one as stated below.
+ * We need to make sure the hw unique we get is valid by:
+ * 1. In case of HUK is used, checking if OTP is hidden (in
+ *    which case only zeros will be returned) or not;
+ * 2. In case of SSK is used, checking if SSK in OTP is
+ *    write_locked (which means a valid key is provisioned)
+ *    or not.
+ *
+ * Maybe tee_get_hw_unique_key() should be exposed as
+ * generic API for getting hw unique key!
+ * We should change the API tee_otp_get_hw_unique_key()
+ * to return error code!
+ */
+static TEE_Result tee_get_hw_unique_key(struct tee_hw_unique_key *hwkey)
+{
+	if (!hwkey)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	tee_otp_get_hw_unique_key(hwkey);
+
+	return TEE_SUCCESS;
+}
+
+static TEE_Result tee_rpmb_key_gen(uint16_t dev_id __unused,
+				   uint8_t *key, uint32_t len)
+{
+	TEE_Result res;
+	struct tee_hw_unique_key hwkey;
+	uint8_t *ctx = NULL;
+
+	if (!key || RPMB_KEY_MAC_SIZE != len) {
+		res = TEE_ERROR_BAD_PARAMETERS;
+		goto out;
+	}
+
+	IMSG("RPMB: Using generated key");
+	res = tee_get_hw_unique_key(&hwkey);
+	if (res != TEE_SUCCESS)
+		goto out;
+
+	ctx = malloc(rpmb_ctx->hash_ctx_size);
+	if (!ctx) {
+		res = TEE_ERROR_OUT_OF_MEMORY;
+		goto out;
+	}
+
+	res = crypto_ops.mac.init(ctx, TEE_ALG_HMAC_SHA256, hwkey.data,
+				  HW_UNIQUE_KEY_LENGTH);
+	if (res != TEE_SUCCESS)
+		goto out;
+
+	res = crypto_ops.mac.update(ctx, TEE_ALG_HMAC_SHA256,
+				    (uint8_t *)rpmb_ctx->cid,
+				    RPMB_EMMC_CID_SIZE);
+	if (res != TEE_SUCCESS)
+		goto out;
+
+	res = crypto_ops.mac.final(ctx, TEE_ALG_HMAC_SHA256, key, len);
+
+out:
+	free(ctx);
+	return res;
+}
+
+#endif /* !CFG_RPMB_TESTKEY */
+
+static void u32_to_bytes(uint32_t u32, uint8_t *bytes)
+{
+	*bytes = (uint8_t) (u32 >> 24);
+	*(bytes + 1) = (uint8_t) (u32 >> 16);
+	*(bytes + 2) = (uint8_t) (u32 >> 8);
+	*(bytes + 3) = (uint8_t) u32;
+}
+
+static void bytes_to_u32(uint8_t *bytes, uint32_t *u32)
+{
+	*u32 = (uint32_t) ((*(bytes) << 24) +
+			   (*(bytes + 1) << 16) +
+			   (*(bytes + 2) << 8) + (*(bytes + 3)));
+}
+
+static void u16_to_bytes(uint16_t u16, uint8_t *bytes)
+{
+	*bytes = (uint8_t) (u16 >> 8);
+	*(bytes + 1) = (uint8_t) u16;
+}
+
+static void bytes_to_u16(uint8_t *bytes, uint16_t *u16)
+{
+	*u16 = (uint16_t) ((*bytes << 8) + *(bytes + 1));
+}
+
+static TEE_Result tee_rpmb_mac_calc(uint8_t *mac, uint32_t macsize,
+				    uint8_t *key, uint32_t keysize,
+				    struct rpmb_data_frame *datafrms,
+				    uint16_t blkcnt)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	int i;
+	uint8_t *ctx = NULL;
+
+	if (!mac || !key || !datafrms)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	ctx = malloc(rpmb_ctx->hash_ctx_size);
+	if (!ctx)
+		return TEE_ERROR_OUT_OF_MEMORY;
+
+	res = crypto_ops.mac.init(ctx, TEE_ALG_HMAC_SHA256, key, keysize);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	for (i = 0; i < blkcnt; i++) {
+		res = crypto_ops.mac.update(ctx, TEE_ALG_HMAC_SHA256,
+					  datafrms[i].data,
+					  RPMB_MAC_PROTECT_DATA_SIZE);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+	}
+
+	res = crypto_ops.mac.final(ctx, TEE_ALG_HMAC_SHA256, mac, macsize);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	free(ctx);
+	return res;
+}
+
+struct tee_rpmb_mem {
+	paddr_t phreq;
+	uint64_t phreq_cookie;
+	paddr_t phresp;
+	uint64_t phresp_cookie;
+	size_t req_size;
+	size_t resp_size;
+};
+
+static void tee_rpmb_free(struct tee_rpmb_mem *mem)
+{
+	if (!mem)
+		return;
+
+	thread_rpc_free_payload(mem->phreq_cookie);
+	thread_rpc_free_payload(mem->phresp_cookie);
+	mem->phreq = 0;
+	mem->phreq_cookie = 0;
+	mem->phresp = 0;
+	mem->phresp_cookie = 0;
+}
+
+
+static TEE_Result tee_rpmb_alloc(size_t req_size, size_t resp_size,
+		struct tee_rpmb_mem *mem, void **req, void **resp)
+{
+	TEE_Result res = TEE_SUCCESS;
+	size_t req_s = ROUNDUP(req_size, sizeof(uint32_t));
+	size_t resp_s = ROUNDUP(resp_size, sizeof(uint32_t));
+
+	if (!mem)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	memset(mem, 0, sizeof(*mem));
+	thread_rpc_alloc_payload(req_s, &mem->phreq, &mem->phreq_cookie);
+	thread_rpc_alloc_payload(resp_s, &mem->phresp, &mem->phresp_cookie);
+	if (!mem->phreq || !mem->phresp) {
+		res = TEE_ERROR_OUT_OF_MEMORY;
+		goto out;
+	}
+
+	*req = phys_to_virt(mem->phreq, MEM_AREA_NSEC_SHM);
+	*resp = phys_to_virt(mem->phresp, MEM_AREA_NSEC_SHM);
+	if (!*req || !*resp) {
+		res = TEE_ERROR_GENERIC;
+		goto out;
+	}
+
+	mem->req_size = req_size;
+	mem->resp_size = resp_size;
+
+out:
+	if (res != TEE_SUCCESS)
+		tee_rpmb_free(mem);
+	return res;
+}
+
+static TEE_Result tee_rpmb_invoke(struct tee_rpmb_mem *mem)
+{
+	struct optee_msg_param params[2];
+
+	memset(params, 0, sizeof(params));
+	params[0].attr = OPTEE_MSG_ATTR_TYPE_TMEM_INPUT;
+	params[1].attr = OPTEE_MSG_ATTR_TYPE_TMEM_OUTPUT;
+	params[0].u.tmem.buf_ptr = mem->phreq;
+	params[0].u.tmem.size = mem->req_size;
+	params[0].u.tmem.shm_ref = mem->phreq_cookie;
+	params[1].u.tmem.buf_ptr = mem->phresp;
+	params[1].u.tmem.size = mem->resp_size;
+	params[1].u.tmem.shm_ref = mem->phresp_cookie;
+
+	return thread_rpc_cmd(OPTEE_MSG_RPC_CMD_RPMB, 2, params);
+}
+
+static bool is_zero(const uint8_t *buf, size_t size)
+{
+	size_t i;
+
+	for (i = 0; i < size; i++)
+		if (buf[i])
+			return false;
+	return true;
+}
+
+#ifdef CFG_ENC_FS
+static TEE_Result encrypt_block(uint8_t *out, const uint8_t *in,
+				uint16_t blk_idx, const uint8_t *fek)
+{
+	return tee_fs_crypt_block(out, in, RPMB_DATA_SIZE, blk_idx, fek,
+				  TEE_MODE_ENCRYPT);
+}
+
+static TEE_Result decrypt_block(uint8_t *out, const uint8_t *in,
+				uint16_t blk_idx, const uint8_t *fek)
+{
+	return tee_fs_crypt_block(out, in, RPMB_DATA_SIZE, blk_idx, fek,
+				  TEE_MODE_DECRYPT);
+}
+#endif /* CFG_ENC_FS */
+
+/* Decrypt/copy at most one block of data */
+static TEE_Result decrypt(uint8_t *out, const struct rpmb_data_frame *frm,
+			  size_t size, size_t offset,
+			  uint16_t blk_idx __maybe_unused, const uint8_t *fek)
+{
+	uint8_t *tmp __maybe_unused;
+
+	TEE_ASSERT(size + offset <= RPMB_DATA_SIZE);
+
+	if (!fek) {
+		/* Block is not encrypted (not a file data block) */
+		memcpy(out, frm->data + offset, size);
+	} else if (is_zero(fek, TEE_FS_KM_FEK_SIZE)) {
+		/*
+		 * The file was created with encryption disabled
+		 * (CFG_ENC_FS=n)
+		 */
+#ifdef CFG_ENC_FS
+		return TEE_ERROR_SECURITY;
+#else
+		memcpy(out, frm->data + offset, size);
+#endif
+	} else {
+		/* Block is encrypted */
+#ifdef CFG_ENC_FS
+		if (size < RPMB_DATA_SIZE) {
+			/*
+			 * Since output buffer is not large enough to hold one
+			 * block we must allocate a temporary buffer.
+			 */
+			tmp = malloc(RPMB_DATA_SIZE);
+			if (!tmp)
+				return TEE_ERROR_OUT_OF_MEMORY;
+			decrypt_block(tmp, frm->data, blk_idx, fek);
+			memcpy(out, tmp + offset, size);
+			free(tmp);
+		} else {
+			TEE_ASSERT(!offset);
+			decrypt_block(out, frm->data, blk_idx, fek);
+		}
+#else
+		return TEE_ERROR_SECURITY;
+#endif
+	}
+
+	return TEE_SUCCESS;
+}
+
+static TEE_Result tee_rpmb_req_pack(struct rpmb_req *req,
+				    struct rpmb_raw_data *rawdata,
+				    uint16_t nbr_frms, uint16_t dev_id,
+				    const uint8_t *fek __unused)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	int i;
+	struct rpmb_data_frame *datafrm;
+
+	if (!req || !rawdata || !nbr_frms)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	/*
+	 * Check write blockcount is not bigger than reliable write
+	 * blockcount.
+	 */
+	if ((rawdata->msg_type == RPMB_MSG_TYPE_REQ_AUTH_DATA_WRITE) &&
+	    (nbr_frms > rpmb_ctx->rel_wr_blkcnt)) {
+		DMSG("wr_blkcnt(%d) > rel_wr_blkcnt(%d)", nbr_frms,
+		     rpmb_ctx->rel_wr_blkcnt);
+		return TEE_ERROR_GENERIC;
+	}
+
+	req->cmd = RPMB_CMD_DATA_REQ;
+	req->dev_id = dev_id;
+
+	/* Allocate memory for construct all data packets and calculate MAC. */
+	datafrm = calloc(nbr_frms, RPMB_DATA_FRAME_SIZE);
+	if (!datafrm)
+		return TEE_ERROR_OUT_OF_MEMORY;
+
+	for (i = 0; i < nbr_frms; i++) {
+		u16_to_bytes(rawdata->msg_type, datafrm[i].msg_type);
+
+		if (rawdata->block_count)
+			u16_to_bytes(*rawdata->block_count,
+				     datafrm[i].block_count);
+
+		if (rawdata->blk_idx) {
+			/* Check the block index is within range. */
+			if ((*rawdata->blk_idx + nbr_frms) >
+			    rpmb_ctx->max_blk_idx) {
+				res = TEE_ERROR_GENERIC;
+				goto func_exit;
+			}
+			u16_to_bytes(*rawdata->blk_idx, datafrm[i].address);
+		}
+
+		if (rawdata->write_counter)
+			u32_to_bytes(*rawdata->write_counter,
+				     datafrm[i].write_counter);
+
+		if (rawdata->nonce)
+			memcpy(datafrm[i].nonce, rawdata->nonce,
+			       RPMB_NONCE_SIZE);
+
+		if (rawdata->data) {
+#ifdef CFG_ENC_FS
+			if (fek)
+				encrypt_block(datafrm[i].data,
+					rawdata->data + (i * RPMB_DATA_SIZE),
+					*rawdata->blk_idx + i, fek);
+			else
+#endif
+				memcpy(datafrm[i].data,
+				       rawdata->data + (i * RPMB_DATA_SIZE),
+				       RPMB_DATA_SIZE);
+		}
+	}
+
+	if (rawdata->key_mac) {
+		if (rawdata->msg_type == RPMB_MSG_TYPE_REQ_AUTH_DATA_WRITE) {
+			res =
+			    tee_rpmb_mac_calc(rawdata->key_mac,
+					      RPMB_KEY_MAC_SIZE, rpmb_ctx->key,
+					      RPMB_KEY_MAC_SIZE, datafrm,
+					      nbr_frms);
+			if (res != TEE_SUCCESS)
+				goto func_exit;
+		}
+		memcpy(datafrm[nbr_frms - 1].key_mac,
+		       rawdata->key_mac, RPMB_KEY_MAC_SIZE);
+	}
+
+	memcpy(TEE_RPMB_REQ_DATA(req), datafrm,
+	       nbr_frms * RPMB_DATA_FRAME_SIZE);
+
+#ifdef CFG_RPMB_FS_DEBUG_DATA
+	for (i = 0; i < nbr_frms; i++) {
+		DMSG("Dumping data frame %d:", i);
+		DHEXDUMP((uint8_t *)&datafrm[i] + RPMB_STUFF_DATA_SIZE,
+			 512 - RPMB_STUFF_DATA_SIZE);
+	}
+#endif
+
+	res = TEE_SUCCESS;
+func_exit:
+	free(datafrm);
+	return res;
+}
+
+static TEE_Result data_cpy_mac_calc_1b(struct rpmb_raw_data *rawdata,
+				       struct rpmb_data_frame *frm,
+				       uint8_t *fek)
+{
+	TEE_Result res;
+	uint8_t *data;
+	uint16_t idx;
+
+	if (rawdata->len + rawdata->byte_offset > RPMB_DATA_SIZE)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	res = tee_rpmb_mac_calc(rawdata->key_mac, RPMB_KEY_MAC_SIZE,
+				rpmb_ctx->key, RPMB_KEY_MAC_SIZE, frm, 1);
+	if (res != TEE_SUCCESS)
+		return res;
+
+	data = rawdata->data;
+	bytes_to_u16(frm->address, &idx);
+
+	res = decrypt(data, frm, rawdata->len, rawdata->byte_offset, idx, fek);
+	return res;
+}
+
+static TEE_Result tee_rpmb_data_cpy_mac_calc(struct rpmb_data_frame *datafrm,
+					     struct rpmb_raw_data *rawdata,
+					     uint16_t nbr_frms,
+					     struct rpmb_data_frame *lastfrm,
+					     uint8_t *fek)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	int i;
+	uint8_t *ctx = NULL;
+	uint16_t offset;
+	uint32_t size;
+	uint8_t *data;
+	uint16_t start_idx;
+	struct rpmb_data_frame localfrm;
+
+	if (!datafrm || !rawdata || !nbr_frms || !lastfrm)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	if (nbr_frms == 1)
+		return data_cpy_mac_calc_1b(rawdata, lastfrm, fek);
+
+	/* nbr_frms > 1 */
+
+	data = rawdata->data;
+
+	ctx = malloc(rpmb_ctx->hash_ctx_size);
+	if (!ctx) {
+		res = TEE_ERROR_OUT_OF_MEMORY;
+		goto func_exit;
+	}
+
+	res = crypto_ops.mac.init(ctx, TEE_ALG_HMAC_SHA256, rpmb_ctx->key,
+				  RPMB_KEY_MAC_SIZE);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	/*
+	 * Note: JEDEC JESD84-B51: "In every packet the address is the start
+	 * address of the full access (not address of the individual half a
+	 * sector)"
+	 */
+	bytes_to_u16(lastfrm->address, &start_idx);
+
+	for (i = 0; i < (nbr_frms - 1); i++) {
+
+		/*
+		 * By working on a local copy of the RPMB frame, we ensure that
+		 * the data can not be modified after the MAC is computed but
+		 * before the payload is decrypted/copied to the output buffer.
+		 */
+		memcpy(&localfrm, &datafrm[i], RPMB_DATA_FRAME_SIZE);
+
+		res = crypto_ops.mac.update(ctx, TEE_ALG_HMAC_SHA256,
+					    localfrm.data,
+					    RPMB_MAC_PROTECT_DATA_SIZE);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+
+		if (i == 0) {
+			/* First block */
+			offset = rawdata->byte_offset;
+			size = RPMB_DATA_SIZE - offset;
+		} else {
+			/* Middle blocks */
+			size = RPMB_DATA_SIZE;
+			offset = 0;
+		}
+
+		res = decrypt(data, &localfrm, size, offset, start_idx + i,
+			      fek);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+
+		data += size;
+	}
+
+	/* Last block */
+	size = (rawdata->len + rawdata->byte_offset) % RPMB_DATA_SIZE;
+	if (size == 0)
+		size = RPMB_DATA_SIZE;
+	res = decrypt(data, lastfrm, size, 0, start_idx + nbr_frms - 1, fek);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	/* Update MAC against the last block */
+	res = crypto_ops.mac.update(ctx, TEE_ALG_HMAC_SHA256, lastfrm->data,
+				    RPMB_MAC_PROTECT_DATA_SIZE);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = crypto_ops.mac.final(ctx, TEE_ALG_HMAC_SHA256, rawdata->key_mac,
+				   RPMB_KEY_MAC_SIZE);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	free(ctx);
+	return res;
+}
+
+static TEE_Result tee_rpmb_resp_unpack_verify(struct rpmb_data_frame *datafrm,
+					      struct rpmb_raw_data *rawdata,
+					      uint16_t nbr_frms, uint8_t *fek)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	uint16_t msg_type;
+	uint32_t wr_cnt;
+	uint16_t blk_idx;
+	uint16_t op_result;
+	struct rpmb_data_frame lastfrm;
+
+	if (!datafrm || !rawdata || !nbr_frms)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+#ifdef CFG_RPMB_FS_DEBUG_DATA
+	for (uint32_t i = 0; i < nbr_frms; i++) {
+		DMSG("Dumping data frame %d:", i);
+		DHEXDUMP((uint8_t *)&datafrm[i] + RPMB_STUFF_DATA_SIZE,
+			 512 - RPMB_STUFF_DATA_SIZE);
+	}
+#endif
+
+	/* Make sure the last data packet can't be modified once verified */
+	memcpy(&lastfrm, &datafrm[nbr_frms - 1], RPMB_DATA_FRAME_SIZE);
+
+	/* Handle operation result and translate to TEEC error code. */
+	bytes_to_u16(lastfrm.op_result, &op_result);
+	if (rawdata->op_result)
+		*rawdata->op_result = op_result;
+	if (op_result != RPMB_RESULT_OK)
+		return TEE_ERROR_GENERIC;
+
+	/* Check the response msg_type. */
+	bytes_to_u16(lastfrm.msg_type, &msg_type);
+	if (msg_type != rawdata->msg_type) {
+		DMSG("Unexpected msg_type (0x%04x != 0x%04x)", msg_type,
+		     rawdata->msg_type);
+		return TEE_ERROR_GENERIC;
+	}
+
+	if (rawdata->blk_idx) {
+		bytes_to_u16(lastfrm.address, &blk_idx);
+		if (blk_idx != *rawdata->blk_idx) {
+			DMSG("Unexpected block index");
+			return TEE_ERROR_GENERIC;
+		}
+	}
+
+	if (rawdata->write_counter) {
+		wr_cnt = *rawdata->write_counter;
+		bytes_to_u32(lastfrm.write_counter, rawdata->write_counter);
+		if (msg_type == RPMB_MSG_TYPE_RESP_AUTH_DATA_WRITE) {
+			/* Verify the write counter is incremented by 1 */
+			if (*rawdata->write_counter != wr_cnt + 1) {
+				DMSG("Counter mismatched (0x%04x/0x%04x)",
+				     *rawdata->write_counter, wr_cnt + 1);
+				return TEE_ERROR_SECURITY;
+			}
+			rpmb_ctx->wr_cnt++;
+		}
+	}
+
+	if (rawdata->nonce) {
+		if (buf_compare_ct(rawdata->nonce, lastfrm.nonce,
+				   RPMB_NONCE_SIZE) != 0) {
+			DMSG("Nonce mismatched");
+			return TEE_ERROR_SECURITY;
+		}
+	}
+
+	if (rawdata->key_mac) {
+		if (msg_type == RPMB_MSG_TYPE_RESP_AUTH_DATA_READ) {
+			if (!rawdata->data)
+				return TEE_ERROR_GENERIC;
+
+			res = tee_rpmb_data_cpy_mac_calc(datafrm, rawdata,
+							 nbr_frms, &lastfrm,
+							 fek);
+
+			if (res != TEE_SUCCESS)
+				return res;
+		} else {
+			/*
+			 * There should be only one data frame for
+			 * other msg types.
+			 */
+			if (nbr_frms != 1)
+				return TEE_ERROR_GENERIC;
+
+			res = tee_rpmb_mac_calc(rawdata->key_mac,
+						RPMB_KEY_MAC_SIZE,
+						rpmb_ctx->key,
+						RPMB_KEY_MAC_SIZE,
+						&lastfrm, 1);
+
+			if (res != TEE_SUCCESS)
+				return res;
+		}
+
+#ifndef CFG_RPMB_FS_NO_MAC
+		if (buf_compare_ct(rawdata->key_mac,
+				   (datafrm + nbr_frms - 1)->key_mac,
+				   RPMB_KEY_MAC_SIZE) != 0) {
+			DMSG("MAC mismatched:");
+#ifdef CFG_RPMB_FS_DEBUG_DATA
+			DHEXDUMP((uint8_t *)rawdata->key_mac, 32);
+#endif
+			return TEE_ERROR_SECURITY;
+		}
+#endif /* !CFG_RPMB_FS_NO_MAC */
+	}
+
+	return TEE_SUCCESS;
+}
+
+static TEE_Result tee_rpmb_get_dev_info(uint16_t dev_id,
+					struct rpmb_dev_info *dev_info)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	struct tee_rpmb_mem mem;
+	struct rpmb_dev_info *di;
+	struct rpmb_req *req = NULL;
+	uint8_t *resp = NULL;
+	uint32_t req_size;
+	uint32_t resp_size;
+
+	if (!dev_info)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	req_size = sizeof(struct rpmb_req);
+	resp_size = sizeof(struct rpmb_dev_info);
+	res = tee_rpmb_alloc(req_size, resp_size, &mem,
+			     (void *)&req, (void *)&resp);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	req->cmd = RPMB_CMD_GET_DEV_INFO;
+	req->dev_id = dev_id;
+
+	di = (struct rpmb_dev_info *)resp;
+	di->ret_code = RPMB_CMD_GET_DEV_INFO_RET_ERROR;
+
+	res = tee_rpmb_invoke(&mem);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	if (di->ret_code != RPMB_CMD_GET_DEV_INFO_RET_OK) {
+		res = TEE_ERROR_GENERIC;
+		goto func_exit;
+	}
+
+	memcpy((uint8_t *)dev_info, resp, sizeof(struct rpmb_dev_info));
+
+#ifdef CFG_RPMB_FS_DEBUG_DATA
+	DMSG("Dumping dev_info:");
+	DHEXDUMP((uint8_t *)dev_info, sizeof(struct rpmb_dev_info));
+#endif
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	tee_rpmb_free(&mem);
+	return res;
+}
+
+static TEE_Result tee_rpmb_init_read_wr_cnt(uint16_t dev_id,
+					    uint32_t *wr_cnt,
+					    uint16_t *op_result)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	struct tee_rpmb_mem mem;
+	uint16_t msg_type;
+	uint8_t nonce[RPMB_NONCE_SIZE];
+	uint8_t hmac[RPMB_KEY_MAC_SIZE];
+	struct rpmb_req *req = NULL;
+	struct rpmb_data_frame *resp = NULL;
+	struct rpmb_raw_data rawdata;
+	uint32_t req_size;
+	uint32_t resp_size;
+
+	if (!wr_cnt)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	req_size = sizeof(struct rpmb_req) + RPMB_DATA_FRAME_SIZE;
+	resp_size = RPMB_DATA_FRAME_SIZE;
+	res = tee_rpmb_alloc(req_size, resp_size, &mem,
+			     (void *)&req, (void *)&resp);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = crypto_ops.prng.read(nonce, RPMB_NONCE_SIZE);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_REQ_WRITE_COUNTER_VAL_READ;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+	rawdata.nonce = nonce;
+
+	res = tee_rpmb_req_pack(req, &rawdata, 1, dev_id, NULL);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = tee_rpmb_invoke(&mem);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_RESP_WRITE_COUNTER_VAL_READ;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+	rawdata.op_result = op_result;
+	rawdata.write_counter = wr_cnt;
+	rawdata.nonce = nonce;
+	rawdata.key_mac = hmac;
+
+	res = tee_rpmb_resp_unpack_verify(resp, &rawdata, 1, NULL);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	tee_rpmb_free(&mem);
+	return res;
+}
+
+static TEE_Result tee_rpmb_verify_key_sync_counter(uint16_t dev_id)
+{
+	uint16_t op_result = 0;
+	TEE_Result res = TEE_ERROR_GENERIC;
+
+	res = tee_rpmb_init_read_wr_cnt(dev_id, &rpmb_ctx->wr_cnt,
+					&op_result);
+
+	if (res == TEE_SUCCESS) {
+		rpmb_ctx->key_verified = true;
+		rpmb_ctx->wr_cnt_synced = true;
+	}
+
+	DMSG("Verify key returning 0x%x\n", res);
+	return res;
+}
+
+static TEE_Result tee_rpmb_write_key(uint16_t dev_id)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	struct tee_rpmb_mem mem = { 0 };
+	uint16_t msg_type;
+	struct rpmb_req *req = NULL;
+	struct rpmb_data_frame *resp = NULL;
+	struct rpmb_raw_data rawdata;
+	uint32_t req_size;
+	uint32_t resp_size;
+
+	req_size = sizeof(struct rpmb_req) + RPMB_DATA_FRAME_SIZE;
+	resp_size = RPMB_DATA_FRAME_SIZE;
+	res = tee_rpmb_alloc(req_size, resp_size, &mem,
+			     (void *)&req, (void *)&resp);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_REQ_AUTH_KEY_PROGRAM;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+	rawdata.key_mac = rpmb_ctx->key;
+
+	res = tee_rpmb_req_pack(req, &rawdata, 1, dev_id, NULL);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = tee_rpmb_invoke(&mem);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_RESP_AUTH_KEY_PROGRAM;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+
+	res = tee_rpmb_resp_unpack_verify(resp, &rawdata, 1, NULL);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	tee_rpmb_free(&mem);
+	return res;
+}
+
+/* True when all the required crypto functions are available */
+static bool have_crypto_ops(void)
+{
+	return (crypto_ops.mac.init && crypto_ops.mac.update &&
+		crypto_ops.mac.final && crypto_ops.prng.read);
+}
+
+/* This function must never return TEE_SUCCESS if rpmb_ctx == NULL */
+static TEE_Result tee_rpmb_init(uint16_t dev_id)
+{
+	TEE_Result res = TEE_SUCCESS;
+	struct rpmb_dev_info dev_info;
+
+	if (!have_crypto_ops())
+		return TEE_ERROR_NOT_SUPPORTED;
+
+	if (!rpmb_ctx) {
+		rpmb_ctx = calloc(1, sizeof(struct tee_rpmb_ctx));
+		if (!rpmb_ctx)
+			return TEE_ERROR_OUT_OF_MEMORY;
+	} else if (rpmb_ctx->dev_id != dev_id) {
+		memset(rpmb_ctx, 0x00, sizeof(struct tee_rpmb_ctx));
+	}
+
+	rpmb_ctx->dev_id = dev_id;
+
+	if (!rpmb_ctx->dev_info_synced) {
+		DMSG("RPMB: Syncing device information");
+
+		dev_info.rpmb_size_mult = 0;
+		dev_info.rel_wr_sec_c = 0;
+		res = tee_rpmb_get_dev_info(dev_id, &dev_info);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+
+		DMSG("RPMB: RPMB size is %d*128 KB", dev_info.rpmb_size_mult);
+		DMSG("RPMB: Reliable Write Sector Count is %d",
+		     dev_info.rel_wr_sec_c);
+
+		if (dev_info.rpmb_size_mult == 0) {
+			res = TEE_ERROR_GENERIC;
+			goto func_exit;
+		}
+
+		rpmb_ctx->max_blk_idx = (dev_info.rpmb_size_mult *
+					 RPMB_SIZE_SINGLE / RPMB_DATA_SIZE) - 1;
+
+		memcpy(rpmb_ctx->cid, dev_info.cid, RPMB_EMMC_CID_SIZE);
+
+		if ((rpmb_ctx->hash_ctx_size == 0) &&
+		    (crypto_ops.mac.get_ctx_size(
+			    TEE_ALG_HMAC_SHA256,
+			    &rpmb_ctx->hash_ctx_size))) {
+			rpmb_ctx->hash_ctx_size = 0;
+			res = TEE_ERROR_GENERIC;
+			goto func_exit;
+		}
+
+#ifdef RPMB_DRIVER_MULTIPLE_WRITE_FIXED
+		rpmb_ctx->rel_wr_blkcnt = dev_info.rel_wr_sec_c * 2;
+#else
+		rpmb_ctx->rel_wr_blkcnt = 1;
+#endif
+
+		rpmb_ctx->dev_info_synced = true;
+	}
+
+	if (!rpmb_ctx->key_derived) {
+		DMSG("RPMB INIT: Deriving key");
+
+		res = tee_rpmb_key_gen(dev_id, rpmb_ctx->key,
+				       RPMB_KEY_MAC_SIZE);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+
+		rpmb_ctx->key_derived = true;
+	}
+
+	/* Perform a write counter read to verify if the key is ok. */
+	if (!rpmb_ctx->wr_cnt_synced || !rpmb_ctx->key_verified) {
+		DMSG("RPMB INIT: Verifying Key");
+
+		res = tee_rpmb_verify_key_sync_counter(dev_id);
+		if (res != TEE_SUCCESS && !rpmb_ctx->key_verified) {
+			/*
+			 * Need to write the key here and verify it.
+			 */
+			DMSG("RPMB INIT: Writing Key");
+			res = tee_rpmb_write_key(dev_id);
+			if (res == TEE_SUCCESS) {
+				DMSG("RPMB INIT: Verifying Key");
+				res = tee_rpmb_verify_key_sync_counter(dev_id);
+			}
+		}
+	}
+
+func_exit:
+	return res;
+}
+
+static TEE_Result tee_rpmb_read_unlocked(uint16_t dev_id, uint32_t addr,
+					 uint8_t *data, uint32_t len,
+					 uint8_t *fek)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	struct tee_rpmb_mem mem = { 0 };
+	uint16_t msg_type;
+	uint8_t nonce[RPMB_NONCE_SIZE];
+	uint8_t hmac[RPMB_KEY_MAC_SIZE];
+	struct rpmb_req *req = NULL;
+	struct rpmb_data_frame *resp = NULL;
+	struct rpmb_raw_data rawdata;
+	uint32_t req_size;
+	uint32_t resp_size;
+	uint16_t blk_idx;
+	uint16_t blkcnt;
+	uint8_t byte_offset;
+
+	if (!data || !len)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	blk_idx = addr / RPMB_DATA_SIZE;
+	byte_offset = addr % RPMB_DATA_SIZE;
+
+	blkcnt =
+	    ROUNDUP(len + byte_offset, RPMB_DATA_SIZE) / RPMB_DATA_SIZE;
+	res = tee_rpmb_init(dev_id);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	req_size = sizeof(struct rpmb_req) + RPMB_DATA_FRAME_SIZE;
+	resp_size = RPMB_DATA_FRAME_SIZE * blkcnt;
+	res = tee_rpmb_alloc(req_size, resp_size, &mem,
+			     (void *)&req, (void *)&resp);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_REQ_AUTH_DATA_READ;
+	res = crypto_ops.prng.read(nonce, RPMB_NONCE_SIZE);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+	rawdata.nonce = nonce;
+	rawdata.blk_idx = &blk_idx;
+	res = tee_rpmb_req_pack(req, &rawdata, 1, dev_id, NULL);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	req->block_count = blkcnt;
+
+	DMSG("Read %u block%s at index %u", blkcnt, ((blkcnt > 1) ? "s" : ""),
+	     blk_idx);
+
+	res = tee_rpmb_invoke(&mem);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	msg_type = RPMB_MSG_TYPE_RESP_AUTH_DATA_READ;
+
+	memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+	rawdata.msg_type = msg_type;
+	rawdata.block_count = &blkcnt;
+	rawdata.blk_idx = &blk_idx;
+	rawdata.nonce = nonce;
+	rawdata.key_mac = hmac;
+	rawdata.data = data;
+
+	rawdata.len = len;
+	rawdata.byte_offset = byte_offset;
+
+	res = tee_rpmb_resp_unpack_verify(resp, &rawdata, blkcnt, fek);
+	if (res != TEE_SUCCESS)
+		goto func_exit;
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	tee_rpmb_free(&mem);
+	return res;
+}
+
+static TEE_Result tee_rpmb_write_blk(uint16_t dev_id, uint16_t blk_idx,
+				     const uint8_t *data_blks, uint16_t blkcnt,
+				     const uint8_t *fek)
+{
+	TEE_Result res;
+	struct tee_rpmb_mem mem;
+	uint16_t msg_type;
+	uint32_t wr_cnt;
+	uint8_t hmac[RPMB_KEY_MAC_SIZE];
+	struct rpmb_req *req = NULL;
+	struct rpmb_data_frame *resp = NULL;
+	struct rpmb_raw_data rawdata;
+	uint32_t req_size;
+	uint32_t resp_size;
+	uint32_t nbr_writes;
+	uint16_t tmp_blkcnt;
+	uint16_t tmp_blk_idx;
+	uint16_t i;
+
+	DMSG("Write %u block%s at index %u", blkcnt, ((blkcnt > 1) ? "s" : ""),
+	     blk_idx);
+
+	if (!data_blks || !blkcnt)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	res = tee_rpmb_init(dev_id);
+	if (res != TEE_SUCCESS)
+		return res;
+
+	/*
+	 * We need to split data when block count
+	 * is bigger than reliable block write count.
+	 */
+	if (blkcnt < rpmb_ctx->rel_wr_blkcnt)
+		req_size = sizeof(struct rpmb_req) +
+		    RPMB_DATA_FRAME_SIZE * blkcnt;
+	else
+		req_size = sizeof(struct rpmb_req) +
+		    RPMB_DATA_FRAME_SIZE * rpmb_ctx->rel_wr_blkcnt;
+
+	resp_size = RPMB_DATA_FRAME_SIZE;
+	res = tee_rpmb_alloc(req_size, resp_size, &mem,
+			     (void *)&req, (void *)&resp);
+	if (res != TEE_SUCCESS)
+		return res;
+
+	nbr_writes = blkcnt / rpmb_ctx->rel_wr_blkcnt;
+	if (blkcnt % rpmb_ctx->rel_wr_blkcnt > 0)
+		nbr_writes += 1;
+
+	tmp_blkcnt = rpmb_ctx->rel_wr_blkcnt;
+	tmp_blk_idx = blk_idx;
+	for (i = 0; i < nbr_writes; i++) {
+		/*
+		 * To handle the last write of block count which is
+		 * equal or smaller than reliable write block count.
+		 */
+		if (i == nbr_writes - 1)
+			tmp_blkcnt = blkcnt - rpmb_ctx->rel_wr_blkcnt *
+			    (nbr_writes - 1);
+
+		msg_type = RPMB_MSG_TYPE_REQ_AUTH_DATA_WRITE;
+		wr_cnt = rpmb_ctx->wr_cnt;
+
+		memset(req, 0x00, req_size);
+		memset(resp, 0x00, resp_size);
+
+		memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+		rawdata.msg_type = msg_type;
+		rawdata.block_count = &tmp_blkcnt;
+		rawdata.blk_idx = &tmp_blk_idx;
+		rawdata.write_counter = &wr_cnt;
+		rawdata.key_mac = hmac;
+		rawdata.data = (uint8_t *)data_blks +
+				i * rpmb_ctx->rel_wr_blkcnt * RPMB_DATA_SIZE;
+
+		res = tee_rpmb_req_pack(req, &rawdata, tmp_blkcnt, dev_id,
+					fek);
+		if (res != TEE_SUCCESS)
+			goto out;
+
+		res = tee_rpmb_invoke(&mem);
+		if (res != TEE_SUCCESS) {
+			/*
+			 * To force wr_cnt sync next time, as it might get
+			 * out of sync due to inconsistent operation result!
+			 */
+			rpmb_ctx->wr_cnt_synced = false;
+			goto out;
+		}
+
+		msg_type = RPMB_MSG_TYPE_RESP_AUTH_DATA_WRITE;
+
+		memset(&rawdata, 0x00, sizeof(struct rpmb_raw_data));
+		rawdata.msg_type = msg_type;
+		rawdata.block_count = &tmp_blkcnt;
+		rawdata.blk_idx = &tmp_blk_idx;
+		rawdata.write_counter = &wr_cnt;
+		rawdata.key_mac = hmac;
+
+		res = tee_rpmb_resp_unpack_verify(resp, &rawdata, 1, NULL);
+		if (res != TEE_SUCCESS) {
+			/*
+			 * To force wr_cnt sync next time, as it might get
+			 * out of sync due to inconsistent operation result!
+			 */
+			rpmb_ctx->wr_cnt_synced = false;
+			goto out;
+		}
+
+		tmp_blk_idx += tmp_blkcnt;
+	}
+
+out:
+	tee_rpmb_free(&mem);
+	return res;
+}
+
+/*
+ * Read RPMB data in bytes.
+ *
+ * @dev_id     Device ID of the eMMC device.
+ * @addr       Byte address of data.
+ * @data       Pointer to the data.
+ * @len        Size of data in bytes.
+ * @fek        SSK-encrypted File Encryption Key or NULL.
+ */
+static TEE_Result tee_rpmb_read(uint16_t dev_id, uint32_t addr, uint8_t *data,
+				uint32_t len, uint8_t *fek)
+{
+	TEE_Result res;
+
+	mutex_lock(&rpmb_mutex);
+	res = tee_rpmb_read_unlocked(dev_id, addr, data, len, fek);
+	mutex_unlock(&rpmb_mutex);
+
+	return res;
+}
+
+static bool tee_rpmb_write_is_atomic(uint16_t dev_id __unused, uint32_t addr,
+				     uint32_t len)
+{
+	uint8_t byte_offset = addr % RPMB_DATA_SIZE;
+	uint16_t blkcnt = ROUNDUP(len + byte_offset,
+				  RPMB_DATA_SIZE) / RPMB_DATA_SIZE;
+
+	return (blkcnt <= rpmb_ctx->rel_wr_blkcnt);
+}
+
+/*
+ * Write RPMB data in bytes.
+ *
+ * @dev_id     Device ID of the eMMC device.
+ * @addr       Byte address of data.
+ * @data       Pointer to the data.
+ * @len        Size of data in bytes.
+ * @fek        SSK-encrypted File Encryption Key or NULL.
+ */
+static TEE_Result tee_rpmb_write(uint16_t dev_id, uint32_t addr,
+				 const uint8_t *data, uint32_t len,
+				 uint8_t *fek)
+{
+	TEE_Result res = TEE_ERROR_GENERIC;
+	uint8_t *data_tmp = NULL;
+	uint16_t blk_idx;
+	uint16_t blkcnt;
+	uint8_t byte_offset;
+
+	mutex_lock(&rpmb_mutex);
+
+	blk_idx = addr / RPMB_DATA_SIZE;
+	byte_offset = addr % RPMB_DATA_SIZE;
+
+	blkcnt =
+	    ROUNDUP(len + byte_offset, RPMB_DATA_SIZE) / RPMB_DATA_SIZE;
+
+	if (byte_offset == 0 && (len % RPMB_DATA_SIZE) == 0) {
+		res = tee_rpmb_write_blk(dev_id, blk_idx, data, blkcnt, fek);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+	} else {
+		data_tmp = calloc(blkcnt, RPMB_DATA_SIZE);
+		if (!data_tmp) {
+			res = TEE_ERROR_OUT_OF_MEMORY;
+			goto func_exit;
+		}
+
+		/* Read the complete blocks */
+		res = tee_rpmb_read_unlocked(dev_id, blk_idx * RPMB_DATA_SIZE,
+					     data_tmp,
+					     blkcnt * RPMB_DATA_SIZE, fek);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+
+		/* Partial update of the data blocks */
+		memcpy(data_tmp + byte_offset, data, len);
+
+		res = tee_rpmb_write_blk(dev_id, blk_idx, data_tmp, blkcnt,
+					 fek);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+	}
+
+	res = TEE_SUCCESS;
+
+func_exit:
+	mutex_unlock(&rpmb_mutex);
+	free(data_tmp);
+	return res;
+}
+
+/*
+ * Read the RPMB write counter.
+ *
+ * @dev_id     Device ID of the eMMC device.
+ * @counter    Pointer to the counter.
+ */
+static TEE_Result tee_rpmb_get_write_counter(uint16_t dev_id,
+					     uint32_t *counter)
+{
+	TEE_Result res = TEE_SUCCESS;
+
+	if (!counter)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	mutex_lock(&rpmb_mutex);
+
+	if (!rpmb_ctx || !rpmb_ctx->wr_cnt_synced) {
+		res = tee_rpmb_init(dev_id);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+	}
+
+	*counter = rpmb_ctx->wr_cnt;
+
+func_exit:
+	mutex_unlock(&rpmb_mutex);
+	return res;
+}
+
+/*
+ * Read the RPMB max block.
+ *
+ * @dev_id     Device ID of the eMMC device.
+ * @counter    Pointer to receive the max block.
+ */
+static TEE_Result tee_rpmb_get_max_block(uint16_t dev_id, uint32_t *max_block)
+{
+	TEE_Result res = TEE_SUCCESS;
+
+	if (!max_block)
+		return TEE_ERROR_BAD_PARAMETERS;
+
+	mutex_lock(&rpmb_mutex);
+
+	if (!rpmb_ctx || !rpmb_ctx->dev_info_synced) {
+		res = tee_rpmb_init(dev_id);
+		if (res != TEE_SUCCESS)
+			goto func_exit;
+	}
+
+	*max_block = rpmb_ctx->max_blk_idx;
+
+func_exit:
+	mutex_unlock(&rpmb_mutex);
+	return res;
+}
+
+/*
+ * End of lower interface to RPMB device
+ */
+
+static TEE_Result get_fat_start_address(uint32_t *addr);
 
 static void dump_fat(void)
 {
@@ -512,16 +1936,6 @@ out:
 }
 
 #ifdef CFG_ENC_FS
-static bool is_zero(const uint8_t *buf, size_t size)
-{
-	size_t i;
-
-	for (i = 0; i < size; i++)
-		if (buf[i])
-			return false;
-	return true;
-}
-
 static TEE_Result generate_fek(struct rpmb_fat_entry *fe)
 {
 	TEE_Result res;
@@ -544,7 +1958,7 @@ static TEE_Result generate_fek(struct rpmb_fat_entry *fe)
 }
 #endif
 
-int tee_rpmb_fs_open(const char *file, int flags, ...)
+static int rpmb_fs_open_internal(const char *file, int flags, ...)
 {
 	int fd = -1;
 	struct rpmb_file_handle *fh = NULL;
@@ -642,6 +2056,7 @@ int tee_rpmb_fs_open(const char *file, int flags, ...)
 	}
 
 	res = TEE_SUCCESS;
+	fh->flags = flags;
 
 out:
 	if (res != TEE_SUCCESS) {
@@ -654,7 +2069,7 @@ out:
 	return fd;
 }
 
-int tee_rpmb_fs_close(int fd)
+static int rpmb_fs_close(int fd)
 {
 	struct rpmb_file_handle *fh;
 
@@ -667,9 +2082,9 @@ int tee_rpmb_fs_close(int fd)
 	return -1;
 }
 
-int tee_rpmb_fs_read(int fd, uint8_t *buf, size_t size)
+static int rpmb_fs_read(TEE_Result *errno, int fd, void *buf, size_t size)
 {
-	TEE_Result res = TEE_ERROR_GENERIC;
+	TEE_Result res;
 	struct rpmb_file_handle *fh;
 	int read_size = -1;
 
@@ -677,43 +2092,43 @@ int tee_rpmb_fs_read(int fd, uint8_t *buf, size_t size)
 		return 0;
 
 	if (!buf) {
-		res = TEE_ERROR_BAD_PARAMETERS;
+		*errno = TEE_ERROR_BAD_PARAMETERS;
 		goto out;
 	}
 
 	fh = handle_lookup(&fs_handle_db, fd);
 	if (!fh) {
-		res = TEE_ERROR_BAD_PARAMETERS;
+		*errno = TEE_ERROR_BAD_PARAMETERS;
 		goto out;
 	}
 	dump_fh(fh);
 
 	res = read_fat(fh, NULL);
-	if (res != TEE_SUCCESS)
+	if (res != TEE_SUCCESS) {
+		*errno = res;
 		goto out;
+	}
 
 	size = MIN(size, fh->fat_entry.data_size - fh->pos);
 	if (size > 0) {
 		res = tee_rpmb_read(CFG_RPMB_FS_DEV_ID,
 				    fh->fat_entry.start_address + fh->pos, buf,
 				    size, fh->fat_entry.fek);
-		if (res != TEE_SUCCESS)
+		if (res != TEE_SUCCESS) {
+			*errno = res;
 			goto out;
+		}
 	}
-
 	read_size = size;
-	res = TEE_SUCCESS;
 
 out:
-	if (res != TEE_SUCCESS)
-		read_size = -1;
-
 	return read_size;
 }
 
-int tee_rpmb_fs_write(int fd, uint8_t *buf, size_t size)
+static int rpmb_fs_write(TEE_Result *errno, int fd, const void *buf,
+			size_t size)
 {
-	TEE_Result res = TEE_ERROR_GENERIC;
+	TEE_Result res;
 	struct rpmb_file_handle *fh;
 	tee_mm_pool_t p;
 	bool pool_result = false;
@@ -821,10 +2236,12 @@ out:
 	if (res == TEE_SUCCESS)
 		return size;
 
+	*errno = res;
 	return -1;
 }
 
-tee_fs_off_t tee_rpmb_fs_lseek(int fd, tee_fs_off_t offset, int whence)
+static tee_fs_off_t rpmb_fs_lseek(TEE_Result *errno, int fd,
+				  tee_fs_off_t offset, int whence)
 {
 	struct rpmb_file_handle *fh;
 	TEE_Result res;
@@ -836,8 +2253,10 @@ tee_fs_off_t tee_rpmb_fs_lseek(int fd, tee_fs_off_t offset, int whence)
 		return TEE_ERROR_BAD_PARAMETERS;
 
 	res = read_fat(fh, NULL);
-	if (res != TEE_SUCCESS)
-		return -1;
+	if (res != TEE_SUCCESS) {
+		*errno = res;
+		goto out;
+	}
 
 	switch (whence) {
 	case TEE_FS_SEEK_SET:
@@ -853,7 +2272,8 @@ tee_fs_off_t tee_rpmb_fs_lseek(int fd, tee_fs_off_t offset, int whence)
 		break;
 
 	default:
-		goto exit;
+		*errno = TEE_ERROR_BAD_PARAMETERS;
+		goto out;
 	}
 
 	if (new_pos < 0)
@@ -861,15 +2281,16 @@ tee_fs_off_t tee_rpmb_fs_lseek(int fd, tee_fs_off_t offset, int whence)
 
 	if (new_pos > TEE_DATA_MAX_POSITION) {
 		EMSG("Position is beyond TEE_DATA_MAX_POSITION");
-		goto exit;
+		*errno = TEE_ERROR_BAD_PARAMETERS;
+		goto out;
 	}
 
 	ret = fh->pos = new_pos;
-exit:
+out:
 	return ret;
 }
 
-TEE_Result tee_rpmb_fs_rm(const char *filename)
+static int rpmb_fs_unlink(const char *filename)
 {
 	TEE_Result res = TEE_ERROR_GENERIC;
 	struct rpmb_file_handle *fh = NULL;
@@ -895,10 +2316,10 @@ TEE_Result tee_rpmb_fs_rm(const char *filename)
 
 out:
 	free(fh);
-	return res;
+	return (res == TEE_SUCCESS ? 0 : -1);
 }
 
-TEE_Result tee_rpmb_fs_rename(const char *old_name, const char *new_name)
+static  int rpmb_fs_rename(const char *old_name, const char *new_name)
 {
 	TEE_Result res = TEE_ERROR_GENERIC;
 	struct rpmb_file_handle *fh_old = NULL;
@@ -952,10 +2373,11 @@ out:
 	free(fh_old);
 	free(fh_new);
 
-	return res;
+	return (res == TEE_SUCCESS ? 0 : -1);
 }
 
-int tee_rpmb_fs_mkdir(const char *path __unused, tee_fs_mode_t mode __unused)
+static int rpmb_fs_mkdir(const char *path __unused,
+			 tee_fs_mode_t mode __unused)
 {
 	/*
 	 * FIXME: mkdir() should really create some entry in the FAT so that
@@ -965,7 +2387,7 @@ int tee_rpmb_fs_mkdir(const char *path __unused, tee_fs_mode_t mode __unused)
 	return 0;
 }
 
-int tee_rpmb_fs_ftruncate(int fd, tee_fs_off_t length)
+static int rpmb_fs_ftruncate(TEE_Result *errno, int fd, tee_fs_off_t length)
 {
 	struct rpmb_file_handle *fh;
 	tee_mm_pool_t p;
@@ -1049,10 +2471,11 @@ out:
 	if (res == TEE_SUCCESS)
 		return 0;
 
+	*errno = res;
 	return -1;
 }
 
-static void tee_rpmb_fs_dir_free(tee_fs_dir *dir)
+static void rpmb_fs_dir_free(struct tee_fs_dir *dir)
 {
 	struct tee_rpmb_fs_dirent *e;
 
@@ -1067,7 +2490,8 @@ static void tee_rpmb_fs_dir_free(tee_fs_dir *dir)
 	}
 }
 
-static TEE_Result tee_rpmb_fs_dir_populate(const char *path, tee_fs_dir *dir)
+static TEE_Result rpmb_fs_dir_populate(const char *path,
+				       struct tee_fs_dir *dir)
 {
 	struct tee_rpmb_fs_dirent *current = NULL;
 	struct rpmb_fat_entry *fat_entries = NULL;
@@ -1158,21 +2582,21 @@ static TEE_Result tee_rpmb_fs_dir_populate(const char *path, tee_fs_dir *dir)
 
 out:
 	if (res != TEE_SUCCESS)
-		tee_rpmb_fs_dir_free(dir);
+		rpmb_fs_dir_free(dir);
 	if (fat_entries)
 		free(fat_entries);
 
 	return res;
 }
 
-static TEE_Result tee_rpmb_fs_opendir_internal(const char *path,
-						tee_fs_dir **dir)
+static TEE_Result rpmb_fs_opendir_internal(const char *path,
+					   struct tee_fs_dir **dir)
 {
 	uint32_t len;
 	uint32_t max_size;
 	char path_local[TEE_RPMB_FS_FILENAME_LENGTH];
 	TEE_Result res = TEE_ERROR_GENERIC;
-	tee_fs_dir *rpmb_dir = NULL;
+	struct tee_fs_dir *rpmb_dir = NULL;
 
 	if (!path || !dir) {
 		res = TEE_ERROR_BAD_PARAMETERS;
@@ -1197,14 +2621,14 @@ static TEE_Result tee_rpmb_fs_opendir_internal(const char *path,
 	if (path_local[len - 1] != '/')
 		path_local[len] = '/';
 
-	rpmb_dir = calloc(1, sizeof(tee_fs_dir));
+	rpmb_dir = calloc(1, sizeof(*rpmb_dir));
 	if (!rpmb_dir) {
 		res = TEE_ERROR_OUT_OF_MEMORY;
 		goto out;
 	}
 	SIMPLEQ_INIT(&rpmb_dir->next);
 
-	res = tee_rpmb_fs_dir_populate(path_local, rpmb_dir);
+	res = rpmb_fs_dir_populate(path_local, rpmb_dir);
 	if (res != TEE_SUCCESS) {
 		free(rpmb_dir);
 		rpmb_dir = NULL;
@@ -1217,12 +2641,12 @@ out:
 	return res;
 }
 
-tee_fs_dir *tee_rpmb_fs_opendir(const char *path)
+static struct tee_fs_dir *rpmb_fs_opendir(const char *path)
 {
-	tee_fs_dir *dir = NULL;
+	struct tee_fs_dir *dir = NULL;
 	TEE_Result res = TEE_ERROR_GENERIC;
 
-	res = tee_rpmb_fs_opendir_internal(path, &dir);
+	res = rpmb_fs_opendir_internal(path, &dir);
 	if (res != TEE_SUCCESS)
 		dir = NULL;
 
@@ -1230,7 +2654,7 @@ tee_fs_dir *tee_rpmb_fs_opendir(const char *path)
 }
 
 
-struct tee_fs_dirent *tee_rpmb_fs_readdir(tee_fs_dir *dir)
+static struct tee_fs_dirent *rpmb_fs_readdir(struct tee_fs_dir *dir)
 {
 	if (!dir)
 		return NULL;
@@ -1246,7 +2670,7 @@ struct tee_fs_dirent *tee_rpmb_fs_readdir(tee_fs_dir *dir)
 	return &dir->current->entry;
 }
 
-int tee_rpmb_fs_closedir(tee_fs_dir *dir)
+static int rpmb_fs_closedir(struct tee_fs_dir *dir)
 {
 	TEE_Result res = TEE_ERROR_GENERIC;
 
@@ -1255,7 +2679,7 @@ int tee_rpmb_fs_closedir(tee_fs_dir *dir)
 		goto out;
 	}
 
-	tee_rpmb_fs_dir_free(dir);
+	rpmb_fs_dir_free(dir);
 	free(dir);
 	res = TEE_SUCCESS;
 out:
@@ -1265,16 +2689,16 @@ out:
 	return -1;
 }
 
-int tee_rpmb_fs_rmdir(const char *path)
+static int rpmb_fs_rmdir(const char *path)
 {
-	tee_fs_dir *dir = NULL;
+	struct tee_fs_dir *dir = NULL;
 	TEE_Result res = TEE_ERROR_GENERIC;
 	int ret = -1;
 
 	/* Open the directory anyting other than NO_DATA is a failure */
-	res = tee_rpmb_fs_opendir_internal(path, &dir);
+	res = rpmb_fs_opendir_internal(path, &dir);
 	if (res == TEE_SUCCESS) {
-		tee_rpmb_fs_closedir(dir);
+		rpmb_fs_closedir(dir);
 		ret = -1;
 
 	} else if (res == TEE_ERROR_NO_DATA) {
@@ -1289,8 +2713,7 @@ int tee_rpmb_fs_rmdir(const char *path)
 	return ret;
 }
 
-TEE_Result tee_rpmb_fs_stat(const char *filename,
-			    struct tee_rpmb_fs_stat *stat)
+static int rpmb_fs_stat(const char *filename, struct tee_rpmb_fs_stat *stat)
 {
 	TEE_Result res = TEE_ERROR_GENERIC;
 	struct rpmb_file_handle *fh = NULL;
@@ -1315,10 +2738,10 @@ TEE_Result tee_rpmb_fs_stat(const char *filename,
 
 out:
 	free(fh);
-	return res;
+	return (res == TEE_SUCCESS ? 0 : -1);
 }
 
-int tee_rpmb_fs_access(const char *filename, int mode)
+static int rpmb_fs_access(const char *filename, int mode)
 {
 	struct tee_rpmb_fs_stat stat;
 	TEE_Result res;
@@ -1326,7 +2749,7 @@ int tee_rpmb_fs_access(const char *filename, int mode)
 	/* Mode is currently ignored, this only checks for existence */
 	(void)mode;
 
-	res = tee_rpmb_fs_stat(filename, &stat);
+	res = rpmb_fs_stat(filename, &stat);
 
 	if (res == TEE_SUCCESS)
 		return 0;
@@ -1334,3 +2757,58 @@ int tee_rpmb_fs_access(const char *filename, int mode)
 	return -1;
 }
 
+static int rpmb_fs_open(TEE_Result *errno, const char *file, int flags, ...)
+{
+	int fd = -1;
+	size_t len;
+
+	assert(errno);
+	*errno = TEE_SUCCESS;
+
+	len = strlen(file) + 1;
+	if (len > TEE_FS_NAME_MAX) {
+		*errno = TEE_ERROR_BAD_PARAMETERS;
+		goto exit;
+	}
+
+	/*
+	 * try to open file without O_CREATE flag, if failed try again with
+	 * O_CREATE flag (to distinguish whether it's a new file or not)
+	 */
+	fd = rpmb_fs_open_internal(file, flags & (~TEE_FS_O_CREATE));
+	if (fd < 0) {
+		if (!(flags & TEE_FS_O_CREATE)) {
+			*errno = TEE_ERROR_ITEM_NOT_FOUND;
+			goto exit;
+		}
+
+		fd = rpmb_fs_open_internal(file, flags);
+		/* File has been created */
+	} else {
+		/* File already exists */
+		if ((flags & TEE_FS_O_CREATE) && (flags & TEE_FS_O_EXCL)) {
+			*errno = TEE_ERROR_ACCESS_CONFLICT;
+			rpmb_fs_close(fd);
+		}
+	}
+
+exit:
+	return fd;
+}
+
+struct tee_file_operations tee_file_ops = {
+	.open = rpmb_fs_open,
+	.close = rpmb_fs_close,
+	.read = rpmb_fs_read,
+	.write = rpmb_fs_write,
+	.lseek = rpmb_fs_lseek,
+	.ftruncate = rpmb_fs_ftruncate,
+	.rename = rpmb_fs_rename,
+	.unlink = rpmb_fs_unlink,
+	.mkdir = rpmb_fs_mkdir,
+	.opendir = rpmb_fs_opendir,
+	.closedir = rpmb_fs_closedir,
+	.readdir = rpmb_fs_readdir,
+	.rmdir = rpmb_fs_rmdir,
+	.access = rpmb_fs_access
+};
